Patching AsyncStateMachine into existing Pipeline
Determines that this is not a feasible direction - there seems to be
some significant multi-threading issues with the library's
initialization of its tables. When each ImportTask is submitted to a
ThreadPoolExecutor separately, calls to `lib.albums()` raises an error
relating to that table not being defined. I suspect this is related to
the database connection not yet being fully initialized in that thread.

There were enough issues trying to integrate this into the existing
Pipeline implementation that I wasn't able to reliably measure any
performance differences.

These challenges, along with the general complexity of the change, I
think point to this being a poor tradeoff of complexity vs. performance.