Add progress bar to the convert plugin.
This required changing the control flow from the Beets pipeline API to use a ThreadPoolExecutor, so that the progress bar updates can happen in the main thread of the program.

There is no performance impact. The previous approach spawned N threads, each of which spawned a subprocess per FFmpeg invocation. This approach also spawns N threads, each of which spawns a subprocess per FFmpeg invocation. The difference is solely in how the threads are managed and how items are assigned to threads for work.

Fix shutdown behavior